# Reduce projected columns for UNION