--- a/.gitignore
+++ b/.gitignore
@@ -25,3 +25,6 @@ server/minidumps/
 ivy-*.xml
 concourse/gporca_vars.yml
 xcode_build
+server/*.md_actual
+server/*.md_expected
+