--- a/data/dxl/minidump/CTEWithVolatileFunction.mdp
+++ b/data/dxl/minidump/CTEWithVolatileFunction.mdp
@@ -153,7 +153,7 @@
     <dxl:Plan Id="0" SpaceSize="60">
       <dxl:Sequence>
         <dxl:Properties>
-          <dxl:Cost StartupCost="0" TotalCost="10.214844" Rows="2.000000" Width="8"/>
+          <dxl:Cost StartupCost="0" TotalCost="10.191406" Rows="2.000000" Width="8"/>
         </dxl:Properties>
         <dxl:ProjList>
           <dxl:ProjElem ColId="4" Alias="sum">
@@ -198,7 +198,7 @@
         </dxl:CTEProducer>
         <dxl:GatherMotion InputSegments="0,1" OutputSegments="-1">
           <dxl:Properties>
-            <dxl:Cost StartupCost="0" TotalCost="7.164062" Rows="2.000000" Width="8"/>
+            <dxl:Cost StartupCost="0" TotalCost="7.140625" Rows="2.000000" Width="8"/>
           </dxl:Properties>
           <dxl:ProjList>
             <dxl:ProjElem ColId="4" Alias="sum">
@@ -207,9 +207,9 @@
           </dxl:ProjList>
           <dxl:Filter/>
           <dxl:SortingColumnList/>
-          <dxl:Result>
+          <dxl:Append IsTarget="false" IsZapped="false">
             <dxl:Properties>
-              <dxl:Cost StartupCost="0" TotalCost="6.156250" Rows="2.000000" Width="8"/>
+              <dxl:Cost StartupCost="0" TotalCost="6.132812" Rows="2.000000" Width="8"/>
             </dxl:Properties>
             <dxl:ProjList>
               <dxl:ProjElem ColId="4" Alias="sum">
@@ -217,23 +217,20 @@
               </dxl:ProjElem>
             </dxl:ProjList>
             <dxl:Filter/>
-            <dxl:OneTimeFilter/>
-            <dxl:Append IsTarget="false" IsZapped="false">
+            <dxl:Result>
               <dxl:Properties>
-                <dxl:Cost StartupCost="0" TotalCost="6.156250" Rows="2.000000" Width="8"/>
+                <dxl:Cost StartupCost="0" TotalCost="2.062500" Rows="1.000000" Width="8"/>
               </dxl:Properties>
               <dxl:ProjList>
                 <dxl:ProjElem ColId="4" Alias="sum">
                   <dxl:Ident ColId="4" ColName="sum" TypeMdid="0.1700.1.0"/>
                 </dxl:ProjElem>
-                <dxl:ProjElem ColId="2" Alias="nextval">
-                  <dxl:Ident ColId="2" ColName="nextval" TypeMdid="0.20.1.0"/>
-                </dxl:ProjElem>
               </dxl:ProjList>
               <dxl:Filter/>
+              <dxl:OneTimeFilter/>
               <dxl:Aggregate AggregationStrategy="Hashed" StreamSafe="false">
                 <dxl:Properties>
-                  <dxl:Cost StartupCost="0" TotalCost="2.078125" Rows="1.000000" Width="16"/>
+                  <dxl:Cost StartupCost="0" TotalCost="2.062500" Rows="1.000000" Width="8"/>
                 </dxl:Properties>
                 <dxl:GroupingColumns>
                   <dxl:GroupingColumn ColId="2"/>
@@ -277,59 +274,56 @@
                   </dxl:CTEConsumer>
                 </dxl:RedistributeMotion>
               </dxl:Aggregate>
-              <dxl:Result>
+            </dxl:Result>
+            <dxl:Result>
+              <dxl:Properties>
+                <dxl:Cost StartupCost="0" TotalCost="3.054688" Rows="1.000000" Width="8"/>
+              </dxl:Properties>
+              <dxl:ProjList>
+                <dxl:ProjElem ColId="7" Alias="sum">
+                  <dxl:Ident ColId="7" ColName="sum" TypeMdid="0.1700.1.0"/>
+                </dxl:ProjElem>
+              </dxl:ProjList>
+              <dxl:Filter/>
+              <dxl:OneTimeFilter/>
+              <dxl:RandomMotion InputSegments="-1" OutputSegments="0,1">
                 <dxl:Properties>
-                  <dxl:Cost StartupCost="0" TotalCost="3.062500" Rows="1.000000" Width="16"/>
+                  <dxl:Cost StartupCost="0" TotalCost="2.046875" Rows="1.000000" Width="8"/>
                 </dxl:Properties>
                 <dxl:ProjList>
                   <dxl:ProjElem ColId="7" Alias="sum">
                     <dxl:Ident ColId="7" ColName="sum" TypeMdid="0.1700.1.0"/>
                   </dxl:ProjElem>
-                  <dxl:ProjElem ColId="8" Alias="?column?">
-                    <dxl:ConstValue TypeMdid="0.20.1.0" IsNull="true" IsByValue="true"/>
-                  </dxl:ProjElem>
                 </dxl:ProjList>
                 <dxl:Filter/>
-                <dxl:OneTimeFilter/>
-                <dxl:RandomMotion InputSegments="-1" OutputSegments="0,1">
+                <dxl:SortingColumnList/>
+                <dxl:Aggregate AggregationStrategy="Plain" StreamSafe="false">
                   <dxl:Properties>
-                    <dxl:Cost StartupCost="0" TotalCost="2.046875" Rows="1.000000" Width="8"/>
+                    <dxl:Cost StartupCost="0" TotalCost="1.039062" Rows="1.000000" Width="8"/>
                   </dxl:Properties>
+                  <dxl:GroupingColumns/>
                   <dxl:ProjList>
                     <dxl:ProjElem ColId="7" Alias="sum">
-                      <dxl:Ident ColId="7" ColName="sum" TypeMdid="0.1700.1.0"/>
+                      <dxl:AggFunc AggMdid="0.2107.1.0" AggDistinct="false" AggStage="Normal">
+                        <dxl:Ident ColId="5" ColName="nextval" TypeMdid="0.20.1.0"/>
+                      </dxl:AggFunc>
                     </dxl:ProjElem>
                   </dxl:ProjList>
                   <dxl:Filter/>
-                  <dxl:SortingColumnList/>
-                  <dxl:Aggregate AggregationStrategy="Plain" StreamSafe="false">
+                  <dxl:CTEConsumer CTEId="0" Columns="5">
                     <dxl:Properties>
-                      <dxl:Cost StartupCost="0" TotalCost="1.039062" Rows="1.000000" Width="8"/>
+                      <dxl:Cost StartupCost="0" TotalCost="0.007812" Rows="1.000000" Width="8"/>
                     </dxl:Properties>
-                    <dxl:GroupingColumns/>
                     <dxl:ProjList>
-                      <dxl:ProjElem ColId="7" Alias="sum">
-                        <dxl:AggFunc AggMdid="0.2107.1.0" AggDistinct="false" AggStage="Normal">
-                          <dxl:Ident ColId="5" ColName="nextval" TypeMdid="0.20.1.0"/>
-                        </dxl:AggFunc>
+                      <dxl:ProjElem ColId="5" Alias="nextval">
+                        <dxl:Ident ColId="5" ColName="nextval" TypeMdid="0.20.1.0"/>
                       </dxl:ProjElem>
                     </dxl:ProjList>
-                    <dxl:Filter/>
-                    <dxl:CTEConsumer CTEId="0" Columns="5">
-                      <dxl:Properties>
-                        <dxl:Cost StartupCost="0" TotalCost="0.007812" Rows="1.000000" Width="8"/>
-                      </dxl:Properties>
-                      <dxl:ProjList>
-                        <dxl:ProjElem ColId="5" Alias="nextval">
-                          <dxl:Ident ColId="5" ColName="nextval" TypeMdid="0.20.1.0"/>
-                        </dxl:ProjElem>
-                      </dxl:ProjList>
-                    </dxl:CTEConsumer>
-                  </dxl:Aggregate>
-                </dxl:RandomMotion>
-              </dxl:Result>
-            </dxl:Append>
-          </dxl:Result>
+                  </dxl:CTEConsumer>
+                </dxl:Aggregate>
+              </dxl:RandomMotion>
+            </dxl:Result>
+          </dxl:Append>
         </dxl:GatherMotion>
       </dxl:Sequence>
     </dxl:Plan>