… CMakeLists.txt
… data/dxl/minidump/CannotPullGrpColAboveAgg.mdp
… data/dxl/minidump/DynamicIndexScan-DefaultPartition.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous-EnabledDateConstraint.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous-NoDTS.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous-Overlapping.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous-PartSelectEquality.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous-PartSelectRange.mdp
… data/dxl/minidump/DynamicIndexScan-Heterogenous.mdp
… data/dxl/minidump/DynamicIndexScan-OpenEndedPartitions.mdp
… data/dxl/minidump/Equivalence-class-project-over-LOJ.mdp
… data/dxl/minidump/FOJ-Condition-False.mdp
… data/dxl/minidump/FullOuterJoin.mdp
… data/dxl/minidump/FullOuterJoin2.mdp
… data/dxl/minidump/IndexApply-Heterogeneous-BothSidesPartitioned.mdp
… data/dxl/minidump/IndexApply-Heterogeneous-DTS.mdp
… data/dxl/minidump/IndexApply-Heterogeneous-NoDTS.mdp
… data/dxl/minidump/IndexApply-InnerSelect-Heterogeneous-DTS.mdp
… data/dxl/minidump/LeftOuter2InnerUnionAllAntiSemiJoin.mdp
… data/dxl/minidump/MS-UnionAll-1.mdp
… data/dxl/minidump/MS-UnionAll-2.mdp
… data/dxl/minidump/MS-UnionAll-5.mdp
… data/dxl/minidump/UnionAll.mdp
… data/dxl/minidump/UnionAllWithTruncatedOutput.mdp
… data/dxl/minidump/UnionGbSubquery.mdp
… libgpopt/src/translate/CTranslatorExprToDXL.cpp
… libgpos/include/gpos/common/CDynamicPtrArray.h