When we have a UNION operator that needs to produce only a subset
of the available columns, then also reduce the project lists of the
children of the union. The fix consists of two changes.

First, in method CPhysicalUnionAll::PcrsRequired(), we require only
those columns from the child that are also needed by the parent of
the union operator.

Second, when generating project lists in the expression to DXL transformation,
we use the required plan properties to project only the required columns.
We need to generate n+1 matching project lists, that of the union operator
itself and the project lists of the n children. We first build an array of
integer indexes of the required columns for the union and we then reduce
each of the n+1 lists using this array, ensuring a match.