@zellerh 2019-04-30
@bhuvnesh2703, here is the result of a performance measurement:

51 columns, 10,000,000 rows per table:

| Table format | num tables in union | table/view      | master, execution time in msec | fix, execution time in msec |
| ------------ | ------------------- | --------------- | ------------------------------ | --------------------------- |
| row          | 1                   | table, 51 cols  | 11307                          | 10863                       |
| row          | 3                   | view, 51 cols   | 11701                          | 11637                       |
| row          | 3                   | view, 3 cols    | 11244                          | 11092                       |
| column       | 1                   | table, 51 cols  | 1981                           | 1944                        |
| column       | 3                   | view, 51 cols   | 18471                          | 2247                        |
| column       | 3                   | view, 3 cols    | 1892                           | 2026                        |

Even though the fix applies to both row and column store, we don't see real savings with row-oriented tables, but we do see an improvement for columnar tables. The fix can only improve the measurements above where we have a 3-way union of tables with 51 columns (the second and fifth measurement). All other results should be the same with and without the fix.