@bhuvnesh2703 2019-04-30
> Even though the fix applies to both row and column store, we don't see real savings with row-oriented tables, but we do see an improvement for columnar tables. The fix can only improve the measurements above where we have a 3-way union of tables with 51 columns (the second and fifth measurement). All other results should be the same with and without the fix.

Thanks Hans for the update. With Rows tables, sure i think it doesn't make a difference as we still have to read the complete row, but in case of columnar tables just the specific column files. This is a great improvement from 18471ms to 2247ms.
LGTM.